requirements.txt
fastapi==0.110.0
starlette==0.36.3
jinja2==3.1.3
uvicorn==0.29.0
pandas
scikit-learn
requests
psycopg2-binary
statsmodels
prometheus-fastapi-instrumentator